move *Service classes to logic package
