+++ service-registration/src/test-integration/java/sg/ncl/service/registration/controllers/RegistrationControllerTest.java
@@ -17,7 +17,7 @@ import org.springframework.web.context.WebApplicationContext;
 import sg.ncl.service.authentication.data.jpa.CredentialsEntity;
 import sg.ncl.service.registration.AbstractTest;
 import sg.ncl.service.registration.Util;
-import sg.ncl.service.team.TeamService;
+import sg.ncl.service.team.logic.TeamService;
 import sg.ncl.service.team.data.jpa.TeamEntity;
 import sg.ncl.service.team.serializers.DateTimeDeserializer;
 import sg.ncl.service.team.serializers.DateTimeSerializer;
