+++ service-team/src/main/java/sg/ncl/service/team/web/TeamsController.java
@@ -3,7 +3,7 @@ package sg.ncl.service.team.web;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.*;
-import sg.ncl.service.team.TeamService;
+import sg.ncl.service.team.logic.TeamService;
 import sg.ncl.service.team.data.jpa.TeamEntity;
 
 import javax.inject.Inject;
